[sw/testing] moved mask_rom_test.h
This partially addresses #7976, specifically the item that states:
"mask_rom_test.h should be relocated to a separate subfolder:
sw/device/silicon_creator/testing/".
