+++ sw/device/silicon_creator/lib/base/mock_abs_mmio.h
@@ -6,8 +6,8 @@
 #define OPENTITAN_SW_DEVICE_SILICON_CREATOR_LIB_BASE_MOCK_ABS_MMIO_H_
 
 #include "sw/device/lib/base/testing/mock_mmio_test_utils.h"
-#include "sw/device/lib/testing/mask_rom_test.h"
 #include "sw/device/silicon_creator/lib/base/abs_mmio.h"
+#include "sw/device/silicon_creator/testing/mask_rom_test.h"
 
 namespace mask_rom_test {
 namespace internal {
