+++ sw/device/silicon_creator/lib/epmp_unittest.cc
@@ -9,8 +9,8 @@
 #include "gtest/gtest.h"
 #include "sw/device/lib/base/bitfield.h"
 #include "sw/device/lib/base/csr.h"
-#include "sw/device/lib/base/testing/mock_csr.h"
-#include "sw/device/lib/testing/mask_rom_test.h"
+#include "sw/device/silicon_creator/lib/base/mock_csr.h"
+#include "sw/device/silicon_creator/testing/mask_rom_test.h"
 
 namespace epmp_unittest {
